fix redux form in component
